requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.12.4
flake8==6.1.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
mccabe==0.7.0
packaging==23.2
parsel==1.8.1
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.7
pyflakes==3.1.0
pymongo==4.5.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.0
scrapyrt==0.15.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tldextract==3.6.0
Twisted==22.10.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.24.0.post1
w3lib==2.1.2
zope.interface==6.0
